UNIT_BINDER should support long long ints
Fixes https://gitlab.com/kicad/code/kicad/-/issues/10870

(cherry picked from commit eb1164fc938a631f9846e18a32d7963321893e2b)

+++ common/widgets/unit_binder.cpp
@@ -278,7 +278,7 @@ bool UNIT_BINDER::Validate( double aMin, double aMax, EDA_UNITS aUnits )
 }
 
 
-void UNIT_BINDER::SetValue( int aValue )
+void UNIT_BINDER::SetValue( long long int aValue )
 {
     double value = aValue;
     double displayValue = m_originTransforms.ToDisplay( value, m_coordType );
